+++ src/lib/components/events/CardEvent.svelte
@@ -31,21 +31,21 @@
 
 <a href={`/eventos/${event.slug}`} class="pl-4 pr-4 lg:px-0 group md:pl-0">
   <div
-    class="overflow-hidden w-64 h-fit hover:border-primary rounded-none transition-all md:h-96 ease-in-out group "
+    class="overflow-hidden w-64 h-fit hover:border-primary rounded-none transition-all md:h-96 ease-in-out group relative"
   >
     {#if totalTicketsLeftStudio === 0 || !event.sell}
       <div
-        class="badge bg-zinc-900 border-none rounded-sm absolute top-2 right-2 uppercase tracking-widest"
+        class="badge bg-red-500/90 text-black font-extrabold border-none rounded-sm absolute py-4 uppercase tracking-widest z-10 w-full"
       >
         Adhesión agotada
       </div>
     {/if}
-    <figure class="z-10 drop-shadow-xl ">
+    <figure class="z-10 drop-shadow-xl">
       <img
         width="600"
         height="600"
         loading="lazy"
-        class="object-cover object-top h-72 md:h-96 w-full hover:scale-110 hover:rotate-2 transition-all "
+        class={`object-cover object-top h-72 md:h-96 w-full hover:scale-110 ease-in hover:rotate-2 group transition-all ${totalTicketsLeftStudio === 0 || !event.sell ? 'grayscale lg:hover:grayscale-0 lg:delay-150' : ''}`}
         src={urlForImage(event.poster).height(600).width(600).quality(80).url()}
         alt={event.title}
       />
@@ -71,3 +71,15 @@
     </p>
   </div>
 </a>
+
+
+<style>
+.badge {
+    top: 0;
+    transition: top 0.3s;
+  }
+
+  .group:hover .badge {
+    top: -50px;
+  }
+</style>